Refactor everything out from `core.py` (but keep it for re-exports)
