+++ prometheus_client/openmetrics/parser.py
@@ -4,7 +4,8 @@ from __future__ import unicode_literals
 
 import math
 
-from .. import core
+from ..metrics_core import Metric, METRIC_LABEL_NAME_RE
+from ..samples import Exemplar, Sample, Timestamp
 
 try:
     import StringIO
@@ -68,12 +69,12 @@ def _parse_timestamp(timestamp):
         raise ValueError("Invalid timestamp: {0!r}".format(timestamp))
     try:
         # Simple int.
-        return core.Timestamp(int(timestamp), 0)
+        return Timestamp(int(timestamp), 0)
     except ValueError:
         try:
             # aaaa.bbbb. Nanosecond resolution supported.
             parts = timestamp.split('.', 1)
-            return core.Timestamp(int(parts[0]), int(parts[1][:9].ljust(9, "0")))
+            return Timestamp(int(parts[0]), int(parts[1][:9].ljust(9, "0")))
         except ValueError:
             # Float.
             ts = float(timestamp)
@@ -110,7 +111,7 @@ def _parse_labels(it, text):
             if char == '\\':
                 state = 'labelvalueslash'
             elif char == '"':
-                if not core._METRIC_LABEL_NAME_RE.match(''.join(labelname)):
+                if not METRIC_LABEL_NAME_RE.match(''.join(labelname)):
                     raise ValueError("Invalid line: " + text)
                 labels[''.join(labelname)] = ''.join(labelvalue)
                 labelname = []
@@ -224,13 +225,13 @@ def _parse_sample(text):
         exemplar_length = sum([len(k) + len(v) + 3 for k, v in exemplar_labels.items()]) + 2
         if exemplar_length > 64:
             raise ValueError("Exmplar labels are too long: " + text)
-        exemplar = core.Exemplar(
+        exemplar = Exemplar(
             exemplar_labels,
             _parse_value(exemplar_value),
             _parse_timestamp(exemplar_timestamp),
         )
 
-    return core.Sample(''.join(name), labels, val, ts, exemplar)
+    return Sample(''.join(name), labels, val, ts, exemplar)
 
 
 def _group_for_sample(sample, name, typ):
@@ -295,7 +296,7 @@ def text_fd_to_metric_families(fd):
     so successful parsing does not imply that the parsed
     text meets the specification.
 
-    Yields core.Metric's.
+    Yields Metric's.
     """
     name = None
     allowed_names = []
@@ -319,7 +320,7 @@ def text_fd_to_metric_families(fd):
             raise ValueError("Units not allowed for this metric type: " + name)
         if typ in ['histogram', 'gaugehistogram']:
             _check_histogram(samples, name)
-        metric = core.Metric(name, documentation, typ, unit)
+        metric = Metric(name, documentation, typ, unit)
         # TODO: check labelvalues are valid utf8
         metric.samples = samples
         return metric
