+++ prometheus_client/process_collector.py
@@ -4,7 +4,8 @@ from __future__ import unicode_literals
 
 import os
 
-from . import core
+from .metrics_core import CounterMetricFamily, GaugeMetricFamily
+from .registry import REGISTRY
 
 try:
     import resource
@@ -17,7 +18,7 @@ except ImportError:
 class ProcessCollector(object):
     """Collector for Standard Exports such as cpu and memory."""
 
-    def __init__(self, namespace='', pid=lambda: 'self', proc='/proc', registry=core.REGISTRY):
+    def __init__(self, namespace='', pid=lambda: 'self', proc='/proc', registry=REGISTRY):
         self._namespace = namespace
         self._pid = pid
         self._proc = proc
@@ -57,17 +58,17 @@ class ProcessCollector(object):
             with open(os.path.join(pid, 'stat'), 'rb') as stat:
                 parts = (stat.read().split(b')')[-1].split())
 
-            vmem = core.GaugeMetricFamily(self._prefix + 'virtual_memory_bytes',
+            vmem = GaugeMetricFamily(self._prefix + 'virtual_memory_bytes',
                                           'Virtual memory size in bytes.', value=float(parts[20]))
-            rss = core.GaugeMetricFamily(self._prefix + 'resident_memory_bytes', 'Resident memory size in bytes.',
+            rss = GaugeMetricFamily(self._prefix + 'resident_memory_bytes', 'Resident memory size in bytes.',
                                          value=float(parts[21]) * _PAGESIZE)
             start_time_secs = float(parts[19]) / self._ticks
-            start_time = core.GaugeMetricFamily(self._prefix + 'start_time_seconds',
+            start_time = GaugeMetricFamily(self._prefix + 'start_time_seconds',
                                                 'Start time of the process since unix epoch in seconds.',
                                                 value=start_time_secs + self._btime)
             utime = float(parts[11]) / self._ticks
             stime = float(parts[12]) / self._ticks
-            cpu = core.CounterMetricFamily(self._prefix + 'cpu_seconds_total',
+            cpu = CounterMetricFamily(self._prefix + 'cpu_seconds_total',
                                            'Total user and system CPU time spent in seconds.',
                                            value=utime + stime)
             result.extend([vmem, rss, start_time, cpu])
@@ -78,11 +79,11 @@ class ProcessCollector(object):
             with open(os.path.join(pid, 'limits'), 'rb') as limits:
                 for line in limits:
                     if line.startswith(b'Max open file'):
-                        max_fds = core.GaugeMetricFamily(self._prefix + 'max_fds',
+                        max_fds = GaugeMetricFamily(self._prefix + 'max_fds',
                                                          'Maximum number of open file descriptors.',
                                                          value=float(line.split()[3]))
                         break
-            open_fds = core.GaugeMetricFamily(self._prefix + 'open_fds',
+            open_fds = GaugeMetricFamily(self._prefix + 'open_fds',
                                               'Number of open file descriptors.',
                                               len(os.listdir(os.path.join(pid, 'fd'))))
             result.extend([open_fds, max_fds])
